requirements.txt
streamlit
face_recognition
google-cloud-documentai
google-cloud-storage
pandas